+++ examples/cli/train/run_training_example.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+python ./download_training_data.py
+sagemaker mxnet train --role-name <your-sagemaker-execution-role>