+++ src/sagemaker/amazon/validation.py
@@ -10,7 +10,6 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-import numbers
 
 
 def gt(minimum):
@@ -45,7 +44,6 @@ def istype(expected):
     def validate(value):
         return isinstance(value, expected)
     return validate
-
 
 isint = istype(int)
 isbool = istype(bool)