+++ tests/unit/test_hyperparameter.py
@@ -16,9 +16,9 @@ from sagemaker.amazon.hyperparameter import Hyperparameter
 
 class Test(object):
 
-    blank = Hyperparameter(name="some-name")
+    blank = Hyperparameter(name="some-name", data_type=int)
     elizabeth = Hyperparameter(name='elizabeth')
-    validated = Hyperparameter(name="validated", validate=lambda value: value > 55)
+    validated = Hyperparameter(name="validated", validate=lambda value: value > 55, data_type=int)
 
 
 def test_blank_access():
@@ -55,3 +55,20 @@ def test_validated():
     x.validated = 66
     with pytest.raises(ValueError):
         x.validated = 23
+
+
+def test_data_type():
+    x = Test()
+    x.validated = 66
+    assert type(x.validated) == Test.__dict__["validated"].data_type
+
+
+def test_from_string():
+    x = Test()
+    value = 65
+
+    x.validated = value
+    from_api = str(value)
+
+    x.validated = from_api
+    assert x.validated == value